+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ts-linq/cli
 
+## 1.3.55
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@5.1.0
+  - @ts-linq/migrations@2.8.11
+  - @ts-linq/core@3.5.4
+  - @ts-linq/metadata@4.1.12
+
 ## 1.3.54
 
 ### Patch Changes